dont send empty facts
